requirements.txt
prefect>=3.0
mlflow>=2.15
ragas>=0.2
transformers>=4.44
peft>=0.12
datasets>=3.0
accelerate>=0.34
torch>=2.4
pydantic>=2.9
pyyaml>=6.0
sentencepiece>=0.2
protobuf>=5.28
bitsandbytes>=0.44
